+++ src/main/java/com/github/lmh01/mgt2mt/mod/GameplayFeatureMod.java
@@ -51,7 +51,7 @@ public class GameplayFeatureMod extends AbstractAdvancedDependentMod {
 
     @Override
     public String[] getCompatibleModToolVersions() {
-        return new String[]{"3.0.0-alpha-1", "3.0.0", "3.0.1", MadGamesTycoon2ModTool.VERSION};
+        return new String[]{"3.0.0-alpha-1", "3.0.0", "3.0.1", "3.0.2", MadGamesTycoon2ModTool.VERSION};
     }
 
     @Override