+++ src/main/java/com/github/lmh01/mgt2mt/mod/ThemeMod.java
@@ -32,7 +32,7 @@ public class ThemeMod extends AbstractSimpleDependentMod {
 
     @Override
     public String[] getCompatibleModToolVersions() {
-        return new String[]{"3.0.0-alpha-1", "3.0.0", "3.0.1", MadGamesTycoon2ModTool.VERSION};
+        return new String[]{"3.0.0-alpha-1", "3.0.0", "3.0.1", "3.0.2", MadGamesTycoon2ModTool.VERSION};
     }
 
     @Override