+++ changelog.md
@@ -1,4 +1,9 @@
 # Changelog
+## Coming in [v3.0.2](https://github.com/LMH01/MGT2_Mod_Tool/releases/tag/v3.0.1)
+### Bug fixes
+- "Import selected" button now works properly and the action will no longer fail.
+- One translation key was not found
+
 ## [v3.0.1](https://github.com/LMH01/MGT2_Mod_Tool/releases/tag/v3.0.1) (Latest Version)
 ### Bug fixes
 - Fixed: "Export selected" button now works properly and the action will no longer fail.