+++ changelog.md
@@ -1,7 +1,9 @@
 # Changelog
 ## Coming in [v3.0.2](https://github.com/LMH01/MGT2_Mod_Tool/releases/tag/v3.0.1)
 ### Other
-- Import: When an image file already exists it is now replaced and the import of the mod is not canceled
+- Import
+  - When an image file already exists it is now replaced and the import of the mod is not canceled
+  - The import will no longer be canceled when a folder can not be accessed
 
 ### Bug fixes
 - "Import selected" button now works properly and the action will no longer fail.